feat: move script to m365_powershell
